+++ crates/api/src/reports/community_report/create.rs
@@ -14,7 +14,7 @@ use lemmy_db_schema::{
   traits::{Crud, Reportable},
 };
 use lemmy_db_views::structs::{CommunityReportView, LocalUserView};
-use lemmy_email::send_new_report_email_to_admins;
+use lemmy_email::admin::send_new_report_email_to_admins;
 use lemmy_utils::error::LemmyResult;
 
 pub async fn create_community_report(
